added upcoming bazaars to vndor dashboard and removed unnecessary route
