+++ client/src/App.tsx
@@ -17,7 +17,6 @@ import SportsFacilities from "@/pages/SportsFacilities";
 import MyEvents from "@/pages/MyEvents";
 import WorkshopApprovals from "@/pages/WorkshopApprovals";
 import VendorRequests from "@/pages/VendorRequests";
-import UpcomingBazaars from "@/pages/UpcomingBazaars";
 import NotFound from "@/pages/not-found";
 
 function Router() {
@@ -36,7 +35,6 @@ function Router() {
       <Route path="/my-events" component={MyEvents} />
       <Route path="/approvals/workshops" component={WorkshopApprovals} />
       <Route path="/admin/vendor-requests" component={VendorRequests} />
-      <Route path="/bazaars" component={UpcomingBazaars} />
       <Route component={NotFound} />
     </Switch>
   );
